Split TilesetReader class into another file
